+++ packages/ozone/src/api/index.ts
@@ -23,6 +23,12 @@ import listMembers from './team/listMembers'
 import getConfig from './server/getConfig'
 import chat from './chat'
 import proxied from './proxied'
+import setAddValues from './set/addValues'
+import setGetValues from './set/getValues'
+import querySets from './set/querySets'
+import upsertSet from './set/upsertSet'
+import setDeleteValues from './set/deleteValues'
+import deleteSet from './set/deleteSet'
 import getRepos from './moderation/getRepos'
 
 export * as health from './health'
@@ -54,5 +60,11 @@ export default function (server: Server, ctx: AppContext) {
   chat(server, ctx)
   proxied(server, ctx)
   getConfig(server, ctx)
+  setAddValues(server, ctx)
+  setGetValues(server, ctx)
+  querySets(server, ctx)
+  upsertSet(server, ctx)
+  setDeleteValues(server, ctx)
+  deleteSet(server, ctx)
   return server
 }