+++ packages/internal/handle-resolver-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto-labs/handle-resolver-node
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`80450cbf2`](https://github.com/bluesky-social/atproto/commit/80450cbf2ca27967ee9fe1a5f4bc590b26f1e6b2)]:
+  - @atproto-labs/fetch-node@0.1.3
+
 ## 0.1.5
 
 ### Patch Changes